Add files via upload
Zetech University Anthem
1. Zetech University,

Untold confidence!
Imparted strength, in character.
The center of excellence,

With God we are able to achieve everything.
Zetech! We invent the future!

2.Zetech University,

The pillar of innovation.
Apex of knowledge and competence.
Our asset is research, and science,

With God we are able to achieve everything.
Zetech! We invent the future!

3.Zetech University,

The fountain of enterprise!
We are comrades so adventurous.
Music, arts, sports, our symbol,

With God we are able to achieve everything.
Zetech! We invent the future!